URL changes & fixes (#3)
This PR changes the URL of the git repo for the tutorials from
`https://github.com/blaizot/ramses-tutorials.git` to
`https://github.com/ramses-organisation/ramses-tutorials.git`, in an
effort to base the RAMSES tutorials on the `ramses-organisation`
version. This helps with Issue
https://github.com/ramses-organisation/ramses-tutorials/issues/1 (at
least partially).

In addition, this introduces some Unicode fixes in the "Fundamentals"
tutorial (− converted to -, ≡ converted to `\\equiv`, ∼ converted to
`\\sim`) to ensure that LaTeX compilation works smoothly.

Also corrects some typos in URLs.