Tutorial updates and fixes (#6)
This PR includes small fixes and modifications for the following
tutorials:
- Cosmological simulation
- Idealised galaxy (including a fix for
https://github.com/ramses-organisation/ramses-tutorials/issues/5)
- Turbulent box
These are mostly details added to the text, fixing typos, small changes
to the markdown files to make them prettier, etc.

These three tutorials have been tested and should run fine (except the
turbulent box where there is a bug with recent gcc versions, see
https://github.com/ramses-organisation/ramses/issues/175 for the issue
and https://github.com/ramses-organisation/ramses/pull/176 for a fix).

The "dense core collapse" tutorial needs to be revisited, but the
changes are beyond the scope of this PR.

---------

Co-authored-by: Corentin Cadiou <contact@cphyc.me>